shop

+++ ex/src/main/java/com/example/ex/model/entity/Publisher.java
@@ -17,7 +17,9 @@ public class Publisher {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "publisher_id")
 
-    private Integer id;
+    private Long id;
     private String name;
+//    private boolean is_deleted;
+//    private boolean is_activated;
 
 }
